Allow parameter references in functions
This did not work beforehand because functions and references were
parsed at the same time, but evaluated later. This parsing failed when a
reference was nested in a function.

To solve this, function evaluation *and parsing* are now done after the
whole inventory is available. This means that all references are already
expanded, so neither the function nor the reference parsing need to be
altered.